add: ping_struct.h
