+++ srcs/ft_ping.c
@@ -1,4 +1,5 @@
 #include "ft_ping.h"
+#include "ping_struct.h"
 #include "error.h"
 #include "utils.h"
 #include "print.h"
