+++ srcs/ping_recv.c
@@ -1,4 +1,5 @@
 #include "ft_ping.h"
+#include "ping_struct.h"
 #include "error.h"
 #include <errno.h>
 #include <stdio.h>
